add lighthouse VCs

+++ cleanup.sh
@@ -17,20 +17,27 @@ fi
 # set current namespace
 kubectl config set-context --current --namespace=${CLUSTER_NAME}
 
-# delete validator clients
+# deploy validator clients
 node_index=0
 while [[ $node_index -lt "$CLUSTER_SIZE" ]]
 do
 export NODE_NAME="node$node_index"
 export VC_INDEX="vc$node_index"
+if [[ "$node_index" -le "$CLUSTER_SIZE/2" ]] && [[ $MIX_VCS == "true" ]]
+then
 eval "cat <<EOF
-$(<./templates/teku-vc.yaml)
+$(<./templates/lighthouse-vc.yaml)
 EOF
 " | kubectl delete -f -
+else
+eval "cat <<EOF
+$(<./templates/teku-vc.yaml)
+" | kubectl delete -f -
+fi
 ((node_index=node_index+1))
 done
 
-# delete charon nodes
+# deploy charon nodes
 node_index=0
 while [[ $node_index -lt "$CLUSTER_SIZE" ]]
 do
